Isolate Password actions/reducers
